data/SebastianBitsch/2d-to-3d-transfer-learning/jobspec-cfg/requirements.txt
torch
torchvision
monai
hydra-core
wandb
omegaconf
matplotlib
numpy
nibabel
medpy